added transformer model
