+++ configs/main_config.yaml
@@ -2,7 +2,7 @@
 
 defaults:
   - data: default
-  - model: simple_cnn
+  - model: transformer
   - training: default
   - trainer: default
   - _self_
