+++ configs/model/transformer.yaml
@@ -1,0 +1,16 @@
+# @package _global_.model
+
+# Model type identifier
+type: transformer
+
+# Model architecture parameters
+# kernel_size: 3
+# init_dim: 64
+# depth: 4
+dropout_rate: 0.2
+in_channels: 5
+out_channels: 2
+patch_size: 4
+embed_dim: 128
+depth: 6
+num_heads: 8
