+++ src/model/simple_cnn.py
@@ -1,0 +1,83 @@
+import torch
+import torch.nn as nn
+
+class ResidualBlock(nn.Module):
+    def __init__(self, in_channels, out_channels, kernel_size=3, stride=1):
+        super().__init__()
+        self.conv1 = nn.Conv2d(in_channels, out_channels, kernel_size, stride=stride, padding=kernel_size // 2)
+        self.bn1 = nn.BatchNorm2d(out_channels)
+        self.relu = nn.ReLU(inplace=True)
+        self.conv2 = nn.Conv2d(out_channels, out_channels, kernel_size, padding=kernel_size // 2)
+        self.bn2 = nn.BatchNorm2d(out_channels)
+
+        # Skip connection
+        self.skip = nn.Sequential()
+        if stride != 1 or in_channels != out_channels:
+            self.skip = nn.Sequential(
+                nn.Conv2d(in_channels, out_channels, kernel_size=1, stride=stride), nn.BatchNorm2d(out_channels)
+            )
+
+    def forward(self, x):
+        identity = x
+
+        out = self.conv1(x)
+        out = self.bn1(out)
+        out = self.relu(out)
+
+        out = self.conv2(out)
+        out = self.bn2(out)
+
+        out += self.skip(identity)
+        out = self.relu(out)
+
+        return out
+
+
+class SimpleCNN(nn.Module):
+    def __init__(
+        self,
+        n_input_channels,
+        n_output_channels,
+        kernel_size=3,
+        init_dim=64,
+        depth=4,
+        dropout_rate=0.2,
+    ):
+        super().__init__()
+
+        # Initial convolution to expand channels
+        self.initial = nn.Sequential(
+            nn.Conv2d(n_input_channels, init_dim, kernel_size=kernel_size, padding=kernel_size // 2),
+            nn.BatchNorm2d(init_dim),
+            nn.ReLU(inplace=True),
+        )
+
+        # Residual blocks with increasing feature dimensions
+        self.res_blocks = nn.ModuleList()
+        current_dim = init_dim
+
+        for i in range(depth):
+            out_dim = current_dim * 2 if i < depth - 1 else current_dim
+            self.res_blocks.append(ResidualBlock(current_dim, out_dim))
+            if i < depth - 1:  # Don't double the final layer
+                current_dim *= 2
+
+        # Final prediction layers
+        self.dropout = nn.Dropout2d(dropout_rate)
+        self.final = nn.Sequential(
+            nn.Conv2d(current_dim, current_dim // 2, kernel_size=kernel_size, padding=kernel_size // 2),
+            nn.BatchNorm2d(current_dim // 2),
+            nn.ReLU(inplace=True),
+            nn.Conv2d(current_dim // 2, n_output_channels, kernel_size=1),
+        )
+
+    def forward(self, x):
+        x = self.initial(x)
+
+        for res_block in self.res_blocks:
+            x = res_block(x)
+
+        x = self.dropout(x)
+        x = self.final(x)
+
+        return x
